// src/partners/interfaces.ts
export interface ttmsByHours {
  id: number;
  address?: string;
  pairAddress?: string;
  priceUSD?: string;
  quoteToken?: string;
  createdAt?: number;
  name?: string;
  symbol?: string;
  change24?: string;
  liquidity?: string;
  volume24?: string;
  token?: object;
  cronCount?: number;
  networkId?: number;
  image?: string;
  liquidityTokenSymbol?: string;
  liquidityTokenAddress?: string;
  txnCount24?: number;
  twitterUrl?: string;
  aiScore: number;
  holders: number;

  votesCount24: number;
  absoluteScore: number;
  tokenAgeScore: number;
  liquidityScore: number;
  scoreFromVotes: number;
  scoreFromVolume: number;
  txnCount24Score: number;
  volumeTwoDaysAgo: string;
  holdersCountScore: number;
  scoreFromChange24: number;
  scoreFromVotesFor24h: number;
  votesPercentageFor7d: number;
  votesPercentageFor24h: number;
  volumeChangePercentage: number;
  votersPercentageFor24h: number;
  holdersGrowthPercentage1h: number;
  scoreFromVolumePercentage: number;
  scoreFromVolumeTwoDaysAgo: number;
  holdersGrowthPercentage24h: number;
  scoreFromVotesPercentageFor7d: number;
  scoreFromVotesPercentageFor24h: number;
  scoreHoldersGrowthPercentage1h: number;
  scoreFromVotersPercentageFor24h: number;
  scoreHoldersGrowthPercentage24h: number;
}